+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-18T04:50:35.672Z",
+    "2026-06-19T06:47:32.188Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {